# Speed up nearest-neighbor routines & structure graph generation