## Summary

The `get_nn_info` routines of many nearest-neighbor flavors in `pymatgen.analysis.local_env` recomputed the original periodic image of sites although this information was already stored on the `PeriodicNeighbor` objects (at least it seems so to me - correct me if I'm wrong).
This resulted in O(N^2) instead of O(N) computational complexity (for structures beyond a certain size, almost all time was spent there).

This PR

 * Changes the `get_nn_info` methods to reuse the information from the `PeriodicNeighbor` objects (1st commit)
 * Introduces a faster implementation of `get_all_nn_info` for the `CutoffDictNN` (2nd commit)

The 2nd commit involves some code duplication and is perhaps not absolutely necessary.
It results in further substantial speedup, however.

Example to demonstrate the speedup (adapted from pymatgen test suite)
```python
from pymatgen.analysis.graphs import *
from pymatgen.analysis.local_env import (
    CovalentBondNN,
    CutOffDictNN,
    MinimumDistanceNN,
    MinimumOKeeffeNN,
    OpenBabelNN,
    VoronoiNN,
)

nacl_lattice = Lattice(
    [
        [3.48543625, 0.0, 2.01231756],
        [1.16181208, 3.28610081, 2.01231756],
        [0.0, 0.0, 4.02463512],
    ]
)
nacl = Structure(nacl_lattice, ["Na", "Cl"], [[0, 0, 0], [0.5, 0.5, 0.5]])

nacl.make_supercell([10,10,10])
print(len(nacl))  # 2000 atoms
```

```
# old
In [2]: %timeit -n1 -r1 nacl_graph = Struc
   ...: tureGraph.with_local_env_strategy(
   ...: nacl, CutOffDictNN({("Cl", "Cl"):
   ...: 5.0}))
6min 56s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

# improved get_nn_info
In [6]: %timeit -n1 -r1 nacl_graph = Struc
   ...: tureGraph.with_local_env_strategy(
   ...: nacl, CutOffDictNN({("Cl", "Cl"):
   ...: 5.0}))
4.42 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

# improved get_all_nn_info
In [2]: %timeit -n1 -r1 nacl_graph = Struc
   ...: tureGraph.with_local_env_strategy(
   ...: nacl, CutOffDictNN({("Cl", "Cl"):
   ...: 5.0}))
444 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

I picked 2000 atoms since the original implementation was so slow.
On larger structures, the difference between commit 1 and commit 1+2 becomes larger than 10x, e.g. for 8192 atoms (16x16x16 supercell):
```
In [1]: %run test.py
8192

# improved get_nn_info
In [2]: %timeit -n1 -r1 nacl_graph = Struc
   ...: tureGraph.with_local_env_strategy(
   ...: nacl, CutOffDictNN({("Cl", "Cl"):
   ...: 5.0}))
1min 3s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

# improved get_all_nn_info
In [2]: %timeit -n1 -r1 nacl_graph = Struc
   ...: tureGraph.with_local_env_strategy(
   ...: nacl, CutOffDictNN({("Cl", "Cl"):
   ...: 5.0}))
1.8 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```


## Checklist

Before a pull request can be merged, the following items must be checked:

- [x] Code is in the [standard Python style](https://www.python.org/dev/peps/pep-0008/). The easiest way to handle this
      is to run the following in the **correct sequence** on your local machine. Start with running
      [black](https://black.readthedocs.io/en/stable/index.html) on your new code. This will automatically reformat
      your code to PEP8 conventions and removes most issues. Then run 
      [pycodestyle](https://pycodestyle.readthedocs.io/en/latest/), followed by 
      [flake8](http://flake8.pycqa.org/en/latest/).
- [x] Docstrings have been added in the [Google docstring format](https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html).
      Run [pydocstyle](http://www.pydocstyle.org/en/2.1.1/index.html) on your code.
- [ ] Type annotations are **highly** encouraged. Run [mypy](http://mypy-lang.org/) to type check your code.
- [ ] Tests have been added for any new functionality or bug fixes.
- [ ] All linting and tests pass.

Note that the CI system will run all the above checks. But it will be much more efficient if you already fix most 
errors prior to submitting the PR. It is highly recommended that you use the pre-commit hook provided in the pymatgen 
repository. Simply `cp pre-commit .git/hooks` and a check will be run prior to allowing commits.