@shyuep 2021-09-07
Thanks. Can you please merge the latest pymatgen master to yours? Those linting errors were fixed recently.